+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Reference docs: syntax highlighting for code samples
 - Navigate to _file:lineNumber_ location from [rg-asm](https://rg.atari.org/tools/rg-asm/) console messages
+- Support `rseven` directive
 
 ### Fixed